enable implicit using, move utils to its own folder, extract some classes out from one file, some format changes and syntax sugars
