+++ CSharpModBase/Input/IInputManager.cs
@@ -1,12 +1,9 @@
-using System;
+namespace CSharpModBase.Input;
 
-namespace CSharpModBase.Input
+public interface IInputManager
 {
-    public interface IInputManager
-    {
-        HotKeyItem RegisterKeyBind(Key key, Action action);
+    HotKeyItem RegisterKeyBind(Key key, Action action);
 
-        HotKeyItem RegisterKeyBind(ModifierKeys modifiers, Key key, Action action);
-        HotKeyItem RegisterGamePadBind(GamePadButton button, Action action);
-    }
+    HotKeyItem RegisterKeyBind(ModifierKeys modifiers, Key key, Action action);
+    HotKeyItem RegisterGamePadBind(GamePadButton button, Action action);
 }
