+++ CSharpModBase/Input/ModifierKeys.cs
@@ -1,14 +1,11 @@
-using System;
+namespace CSharpModBase.Input;
 
-namespace CSharpModBase.Input
+[Flags]
+public enum ModifierKeys
 {
-    [Flags]
-    public enum ModifierKeys
-    {
-        None = 0,
-        Alt = 1,
-        Control = 2,
-        Shift = 4,
-        Windows = 8
-    }
+    None = 0,
+    Alt = 1,
+    Control = 2,
+    Shift = 4,
+    Windows = 8
 }
